Simplify layout a bit. Add launch game / status button
