--- internal/detector/event.go
@@ -1,64 +1,0 @@
-package detector
-
-import (
-	"net"
-	"time"
-
-	"github.com/leighmacdonald/bd/internal/store"
-	"github.com/leighmacdonald/steamid/v3/steamid"
-	"github.com/pkg/errors"
-)
-
-const logTimestampFormat = "01/02/2006 - 15:04:05"
-
-// parseTimestamp will convert the source formatted log timestamps into a time.Time value.
-func parseTimestamp(timestamp string) (time.Time, error) {
-	parsedTime, errParse := time.Parse(logTimestampFormat, timestamp)
-	if errParse != nil {
-		return time.Time{}, errors.Wrap(errParse, "Failed to parse timestamp")
-	}
-
-	return parsedTime, nil
-}
-
-type LogEvent struct {
-	Type            EventType
-	Player          string
-	PlayerPing      int
-	PlayerConnected time.Duration
-	Team            store.Team
-	UserID          int
-	PlayerSID       steamid.SID64
-	Victim          string
-	VictimSID       steamid.SID64
-	Message         string
-	Timestamp       time.Time
-	MetaData        string
-	Dead            bool
-	TeamOnly        bool
-}
-
-func (e *LogEvent) ApplyTimestamp(tsString string) error {
-	ts, errTS := parseTimestamp(tsString)
-	if errTS != nil {
-		return errTS
-	}
-
-	e.Timestamp = ts
-
-	return nil
-}
-
-type Event struct {
-	Name  EventType
-	Value any
-}
-
-type Server struct {
-	ServerName string    `json:"server_name"`
-	Addr       net.IP    `json:"-"`
-	Port       uint16    `json:"-"`
-	CurrentMap string    `json:"current_map"`
-	Tags       []string  `json:"-"`
-	LastUpdate time.Time `json:"last_update"`
-}
